fixed errors in Api project
