+++ CodeStyle.ruleset
@@ -1,5 +1,6 @@
 ﻿<RuleSet Name="Rules for PhysEdJournal project" Description="Rule set for StyleCopAnalyzer" ToolsVersion="14.0">
     <Rules AnalyzerId="StyleCop.Analyzers" RuleNamespace="StyleCop.Analyzers">
+        <Rule Id="SA0001" Action="None" />
         <Rule Id="SA1000" Action="None" />
         <Rule Id="SA1009" Action="None" />
         <Rule Id="SA1101" Action="None" />
@@ -8,6 +9,7 @@
         <Rule Id="SA1137" Action="None" />
         <Rule Id="SA1200" Action="None" />
         <Rule Id="SA1201" Action="None" />
+        <Rule Id="SA1202" Action="None" />
         <Rule Id="SA1206" Action="None" />
         <Rule Id="SA1309" Action="None" />
         <Rule Id="SA1500" Action="None" />
