+++ PhysEdJournal.Api/Endpoints/MeEndpoint/Contracts.cs
@@ -2,18 +2,18 @@
 
 public sealed class StudentInfoResponse
 {
-    public double Points { get; init; }
+    public required double Points { get; init; }
 }
 
 public sealed class ProfessorInfoResponse
 {
-    public List<string> Permisions { get; init; }
+    public required List<string> Permisions { get; init; }
 }
 
 public enum UserType
 {
     Student,
-    Professor
+    Professor,
 }
 
 public sealed class MeRequest
